Fix MESSENGER and RUINS interaction (issue #67) (#68)
Fixes 2 issues with our existing MESSENGER implementation:
(1) When removing a Construction from the city, we look to see if it has a MESSENGER that shares a space with it. We updated the played card to say that the MESSENGER doesn't share a space with that Construction anymore, but we pass the const sharedMessenger instead of passing the updated card info
(2) When a Construction is played, we check to see if there are any unpaired messengers. If there are, we pair it to the newly played Construction. However, when a RUINS is played, if there are other eligible constructions to choose from, we want to give the player the option to choose which Construction to pair the MESSENGER with (vs. automatically pairing it with the RUINS)

Creates relocateMessenger and calls it in player.addToCity and player.removeCardFromCity